refactor: restructure routes & dashboard
- Replace old admin dashboard with unified dashboard for admin/employee
- Add role-based access to scan/storage routes
- Consolidate employee scan routes
- Add unified widget map for permissions
- Implement salary & product charts in new layout
