--- src/routes/_authenticated/employee/scan/index.tsx
@@ -1,6 +1,0 @@
-import ScanProduct from '@pages/employee/scan/ScanProduct';
-import { createFileRoute } from '@tanstack/react-router';
-
-export const Route = createFileRoute('/_authenticated/employee/scan/')({
-  component: ScanProduct
-});
