--- src/routes/_authenticated/employee/scan/storage.tsx
@@ -1,6 +1,0 @@
-import { Storage } from '@pages/employee/scan/Storage';
-import { createFileRoute } from '@tanstack/react-router';
-
-export const Route = createFileRoute('/_authenticated/employee/scan/storage')({
-  component: Storage
-});
