+++ src/routes/_authenticated/scan/index.tsx
@@ -1,7 +1,9 @@
+import ScanProduct from '@pages/employee/scan/ScanProduct';
 import { createFileRoute } from '@tanstack/react-router';
 import { requireRole } from '@utils/authUtil';
 
-export const Route = createFileRoute('/_authenticated/employee/scan')({
+export const Route = createFileRoute('/_authenticated/scan/')({
+  component: ScanProduct,
   beforeLoad: async ({ context }) => {
     const { user } = context.authenticated;
     requireRole(user, ['kho']);
